+++ Changelog.md
@@ -1,7 +1,18 @@
-# Changelog v1.0
+# Changelog v1.1
 
 ## Additions
-- Added mossy vertical slabs for all mossy slabs from Nemo's Mossy Blocks
+- Added Mossy Cobbled Deepslate Vertical Slab
+- Added Mossy Deepslate Brick Vertical Slab
+- Added Mossy Deepslate Tile Vertical Slab
+- Added Mossy Tuff Vertical Slab
+- Added Mossy Tuff Brick Vertical Slab
+
+## Fixes
+- Mossy vertical slabs are crafted with normal blocks instead of mossy blocks
+- Wooden mossy vertical slab recipes are unlocked by having normal planks instead of mossy planks
+
+## Changes
+- Put wooden vertical slab recipes in a group
 
 ### Downloads
 [CurseForge](https://curseforge.com/minecraft/mc-mods/nemos-mossy-vertical-slabs) <br>